requirements.txt
torch>=1.12
torchvision
pillow
pandas
scikit-learn
tqdm
